#1048 P1: relocate main.rs inline tests to main_tests.rs
main.rs went from 2,139 raw LOC (1,283 prod / 855 tests) to 1,286 by
moving the bottom `#[cfg(test)] mod tests { ... }` block (line 1284)
into a sibling `main_tests.rs` file, loaded via
`#[cfg(test)] #[path = "main_tests.rs"] mod tests;` appended to main.rs.

Two existing `#[cfg(test)]` decorations stay in place:
- `mod test_zone_ids;` at line 14 (test-only module declaration)
- `fn same_binding_plan(...)` at line 1048 (test-only helper fn)

Pure file-organization split — no test bodies modified, no production
code touched. Test module path `crate::tests` is unchanged.

LOC distribution after:
- main.rs:        1,286 (was 2,139) — under threshold
- main_tests.rs:    857 (new — same content, dedented one level)

cargo test --release: 865 passed, 0 failed.

Closes the LOC-driven portion of #1048. Further structural splits
(server/handlers/state submodules) are deferred — main.rs already
meets the threshold after this PR.
